+++ version/types.go
@@ -1,0 +1,8 @@
+package version
+
+// Info - Binary Version type
+type Info struct {
+	ProgName       string
+	Version        string
+	PackageManager string
+}
